Fix Flutter compilation errors for APK build
- Fixed EmptyState widget to accept optional icon parameter
- Made PaymentHistoryScreen bookingId parameter optional with fallback logic
- Fixed DateTime null safety error in booking_payment_screen.dart
- Fixed num to double conversion in booking_checkout_screen.dart
- Replaced Icons.database with Icons.storage in settings_maintenance.dart

All compilation errors preventing APK build have been resolved.

+++ mobile/lib/screens/payments/booking_checkout_screen.dart
@@ -127,7 +127,7 @@ class _BookingCheckoutScreenState extends ConsumerState<BookingCheckoutScreen> {
                                   const SizedBox(height: 6),
                                   _buildSummaryRow('إجمالي المدفوع', totalPaid, Colors.green),
                                   const SizedBox(height: 6),
-                                  _buildSummaryRow('المتبقي', remainingAmount, remainingAmount <= 0 ? Colors.green : Colors.red),
+                                  _buildSummaryRow('المتبقي', remainingAmount.toDouble(), remainingAmount <= 0 ? Colors.green : Colors.red),
                                 ],
                               ),
                             ),
